--- src/main/java/CancelGPT/exception/task/InvalidTaskException.java
@@ -1,13 +1,0 @@
-package CancelGPT.exception.task;
-
-/**
- * Represents an Exception for invalid task.
- */
-public class InvalidTaskException extends Exception {
-    /**
-     * Initialises InvalidTaskException with message indicating invalid task.
-     */
-    public InvalidTaskException(String message) {
-        super(message);
-    }
-}
