--- src/main/java/CancelGPT/exception/task/MarkTaskInputException.java
@@ -1,13 +1,0 @@
-package CancelGPT.exception.task;
-
-/**
- * Represents an Exception for mark task input.
- */
-public class MarkTaskInputException extends MarkUnmarkTaskInputException {
-    /**
-     * Initialises MarkTaskInputException with message indication mark task input exception.
-     */
-    public MarkTaskInputException() {
-        super("mark");
-    }
-}
