--- src/main/java/CancelGPT/exception/task/TaskDoesNotExist.java
@@ -1,13 +1,0 @@
-package CancelGPT.exception.task;
-
-/**
- * Represents an Exception for task not existing.
- */
-public class TaskDoesNotExist extends Exception {
-    /**
-     * Initialises TaskDoesNotExist with message indicating the task does not exist.
-     */
-    public TaskDoesNotExist() {
-        super("The task you are trying to access does not exist");
-    }
-}
