--- src/main/java/CancelGPT/exception/task/UnmarkTaskInputException.java
@@ -1,13 +1,0 @@
-package CancelGPT.exception.task;
-
-/**
- * Represents an Exception for unmark task input.
- */
-public class UnmarkTaskInputException extends MarkUnmarkTaskInputException {
-    /**
-     * Initialises UnMarkTaskInputException with message indication unmark task input exception.
-     */
-    public UnmarkTaskInputException() {
-        super("unmark");
-    }
-}
