+++ src/main/java/cancelgpt/exception/task/MarkTaskInputException.java
@@ -1,0 +1,13 @@
+package cancelgpt.exception.task;
+
+/**
+ * Represents an Exception for mark task input.
+ */
+public class MarkTaskInputException extends MarkUnmarkTaskInputException {
+    /**
+     * Initialises MarkTaskInputException with message indication mark task input exception.
+     */
+    public MarkTaskInputException() {
+        super("mark");
+    }
+}
